bonito: Switch AUTOSLEEP for CPU_QUIET_GOVERNOR_RUNNABLE
AUTOSLEEP was never originally intended for this device and causes more
trouble than required (unsuccessful suspend attempts in a loop, for example).

On the other hand CPUQUIET provides some ways of ensuring less power waste
when idling with the screen of, by setting the appropriate range of CPUs
to be isolated (not offlined as previously).